fix(TDI-46005): cve bump httpclient to 4.5.13
