--- main/plugins/org.talend.designer.components.libs/libs_src/CouchDBUtil/build.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project name="org.talend.designer.components.libs" default="process" basedir=".">
-	
-	<property name="couchdb.lib.home" value="../../../org.talend.libraries.couchdb/lib" />
-	<property name="apache.lib.home" value="../../../org.talend.libraries.apache.common/lib" />
-	<property name="http.lib.home" value="../../../org.talend.libraries.apache.http/lib" />
-	<!-- #################################################### -->
-
-	<!-- modification 1: config -->
-	<property name="jar.name" value="CouchDBUtils.jar" />
-	<property name="component.name" value="tCouchDBInput" />
-	<property name="author.name" value="jzhao" />
-	
-	<!-- modification 2: compile classpath -->
-	<path id="compile.classpath">
-		<pathelement location="${couchdb.lib.home}/lightcouch-0.0.6.jar" />
-		<pathelement location="${couchdb.lib.home}/gson-2.2.4.jar" />
-		<pathelement location="${apache.lib.home}/commons-codec-1.6.jar" />
-		<pathelement location="${apache.lib.home}/commons-logging-1.1.jar" />
-		<pathelement location="${http.lib.home}/httpclient-4.2.5.jar" />
-		<pathelement location="${http.lib.home}/httpcore-4.2.4.jar" />
-	</path>
-
-	<!-- #################################################### -->	
-	
-
-	<!-- sourcecode and final jar path -->
-	<property name="source.home" value="." />
-	<property name="jar.home" value="${couchdb.lib.home}/${jar.name}" />
-
-	<!-- temp dir for clasee files -->
-	<property name="build.dir" value="../../build" />
-
-	<!-- compile option -->
-	<property name="compile.debug" value="true" />
-	<property name="compile.deprecation" value="false" />
-	<property name="compile.optimize" value="true" />
-
-
-	<target name="process" description="prepare a temp dir">
-		<antcall target="prepare" />
-		<antcall target="compile" />
-		<antcall target="clean" />		
-	</target>
-
-	<target name="prepare" description="prepare a temp dir">
-		<delete dir="${build.dir}" />
-		<mkdir dir="${build.dir}" />
-		<mkdir dir="${build.dir}/classes" />
-	</target>
-
-	<target name="compile" description="Compile Java sources">
-
-		<!-- compile -->
-		<javac srcdir="${source.home}" destdir="${build.dir}/classes" debug="${compile.debug}" deprecation="${compile.deprecation}" optimize="${compile.optimize}">
-			<classpath refid="compile.classpath" />
-		</javac>
-		
-		<!-- include source code -->
-		<copy todir="${build.dir}/classes">
-			<fileset dir="${source.home}">
-				<exclude name="build.xml" />
-			</fileset>
-		</copy>
-
-		<!-- make jar -->
-		<tstamp>
-			<format property="date" pattern="yyyy-MM-dd HH:mm:ss" />
-		</tstamp>
-
-		<jar destfile="${build.dir}/${jar.name}" basedir="${build.dir}/classes">
-			<manifest>
-				<!-- who -->
-				<attribute name="Built-By" value="${author.name}" />
-				<!-- when -->
-				<attribute name="Built-Date" value="${date}"/>  
-				<!-- JDK version -->
-				<attribute name="Created-By" value="${java.version} (${java.vendor})" />
-				<!-- Information about the program itself -->
-				<attribute name="Implementation-Vendor" value="Talend SA" />
-				<attribute name="Implementation-Title" value="${jar.name}" />
-				<attribute name="Implementation-Version" value="1.0" />
-			</manifest>
-		</jar>
-
-		<!-- move jar -->
-		<move file="${build.dir}/${jar.name}" tofile="${jar.home}" />
-
-	</target>
-
-	<target name="clean" description="clean the temp dir">
-		<delete dir="${build.dir}" />
-		<mkdir dir="${build.dir}" />
-	</target>
-</project>
