+++ main/plugins/org.talend.designer.components.libs/libs_src/servicenow/pom.xml
@@ -20,10 +20,19 @@
     <groupId>org.talend.components</groupId>
     <artifactId>components-servicenow</artifactId>
     <name>servicenow</name>
-    <version>1.0.1-20200824</version>
+    <version>1.0.1-20210428</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
         <talend.nexus.url>https://artifacts-oss.talend.com</talend.nexus.url>
+
+        <httpcore.version>4.4.13</httpcore.version>
+        <httpclient.version>4.5.13</httpclient.version>
+        <httpclient-cache.version>4.5.13</httpclient-cache.version>
+        <httpmime.version>4.5.13</httpmime.version>
+        <junit.version>4.13.2</junit.version>
+        <commons-codec.version>1.14</commons-codec.version>
+        <commons-logging.version>1.1.3</commons-logging.version>
     </properties>
     <description>Talend Helper library for ServiceNow components</description>
     <url>
@@ -41,37 +50,37 @@
         <dependency>
             <groupId>commons-codec</groupId>
             <artifactId>commons-codec</artifactId>
-            <version>1.14</version>
+            <version>${commons-codec.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>commons-logging</groupId>
             <artifactId>commons-logging</artifactId>
-            <version>1.1.3</version>
+            <version>${commons-logging.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient</artifactId>
-            <version>4.3.1</version>
+            <version>${httpclient.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient-cache</artifactId>
-            <version>4.3.1</version>
+            <version>${httpclient-cache.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpmime</artifactId>
-            <version>4.3.1</version>
+            <version>${httpmime.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpcore</artifactId>
-            <version>4.3.1</version>
+            <version>${httpcore.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
@@ -83,8 +92,9 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.0</version>
+            <version>${junit.version}</version>
             <type>jar</type>
+            <scope>test</scope>
             <optional>true</optional>
         </dependency>
     </dependencies>
@@ -105,8 +115,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>2.3.2</version>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
             <plugin>
