+++ main/plugins/org.talend.designer.components.libs/libs_src/talend-ws/pom.xml
@@ -4,15 +4,18 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.talend.components</groupId>
 	<artifactId>talend-ws</artifactId>
-	<version>1.0.3-20201222</version>
+	<version>1.0.3-20210428</version>
 	<packaging>jar</packaging>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+		<talend.nexus.url>https://artifacts-oss.talend.com</talend.nexus.url>
+
 		<cxf.version>3.3.7</cxf.version>
 		<odata.version>4.3.0</odata.version>
 		<slf4j.version>1.7.12</slf4j.version>
-		<talend.nexus.url>https://artifacts-oss.talend.com</talend.nexus.url>
+		<httpclient.version>4.5.13</httpclient.version>
 	</properties>
 
 	<distributionManagement>
@@ -49,7 +52,7 @@
 		<dependency>
 			<groupId>org.apache.httpcomponents</groupId>
 			<artifactId>httpclient</artifactId>
-			<version>4.5.12</version>
+			<version>${httpclient.version}</version>
 		</dependency>
 		<!-- https://mvnrepository.com/artifact/org.apache.httpcomponents/httpcore -->
 		<dependency>
@@ -250,6 +253,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
 				<configuration>
 					<source>1.8</source>
 					<target>1.8</target>
